Remove Inheritance and introduce Media Element Player
